fix: Consolidate email valid feature into existing Util class #234
