--- src/Fan/Helpers/RegexUtilities.cs
@@ -1,63 +1,0 @@
-using System;
-using System.Globalization;
-using System.Text.RegularExpressions;
-
-namespace Fan.Helpers
-{
-    /// <summary>
-    /// A util class that uses regular expression to verify if a string is in valid email format.
-    /// </summary>
-    /// <remarks>
-    /// https://docs.microsoft.com/en-us/dotnet/standard/base-types/how-to-verify-that-strings-are-in-valid-email-format
-    /// </remarks>
-    public static class RegexUtilities
-    {
-        public static bool IsValidEmail(string strIn)
-        {
-            if (String.IsNullOrEmpty(strIn))
-                return false;
-
-            // Use IdnMapping class to convert Unicode domain names.
-            try
-            {
-                strIn = Regex.Replace(strIn, @"(@)(.+)$", DomainMapper,
-                                      RegexOptions.None, TimeSpan.FromMilliseconds(200));
-                if(string.IsNullOrEmpty(strIn))
-                    return false;
-            }
-            catch (RegexMatchTimeoutException)
-            {
-                return false;
-            }
-            // Return true if strIn is in valid email format.
-            try
-            {
-                return Regex.IsMatch(strIn,
-                      @"^(?("")("".+?(?<!\\)""@)|(([0-9a-z]((\.(?!\.))|[-!#\$%&'\*\+/=\?\^`\{\}\|~\w])*)(?<=[0-9a-z])@))" +
-                      @"(?(\[)(\[(\d{1,3}\.){3}\d{1,3}\])|(([0-9a-z][-0-9a-z]*[0-9a-z]*\.)+[a-z0-9][\-a-z0-9]{0,22}[a-z0-9]))$",
-                      RegexOptions.IgnoreCase, TimeSpan.FromMilliseconds(250));
-            }
-            catch (RegexMatchTimeoutException)
-            {
-                return false;
-            }
-        }
-
-        private static string DomainMapper(Match match)
-        {
-            // IdnMapping class with default property values.
-            IdnMapping idn = new IdnMapping();
-
-            string domainName = match.Groups[2].Value;
-            try
-            {
-                domainName = idn.GetAscii(domainName);
-            }
-            catch (ArgumentException)
-            {
-                return null; 
-            }
-            return match.Groups[1].Value + domainName;
-        }
-    }
-}
